move logic to base obj structure
