+++ src/client/_converter.py
@@ -1,5 +1,4 @@
 __all__: list = []
-from copy import deepcopy
 import datetime
 from decimal import Decimal
 import typing
@@ -8,18 +7,10 @@ import uuid
 import pkg_resources
 from typing import List, Union
 
-import attr
 import cattr
 from ..util._extendable_enum import ExtendableStrEnum
 
 _CATTRS_VERSION = tuple(map(int, pkg_resources.get_distribution('cattrs').version.split('.')))
-
-if _CATTRS_VERSION < (22, 1, 0):
-    from cattr.gen import is_generic, get_origin, _generate_mapping
-    from cattr._compat import has_with_generic
-else:
-    from cattrs.gen import is_generic, get_origin, _generate_mapping
-    from cattrs._compat import has_with_generic
 
 if _CATTRS_VERSION < (22, 2, 0):
     converter = cattr.Converter()
@@ -28,58 +19,11 @@ else:
 
 converter.register_structure_hook_func(
     lambda type_: hasattr(type_, 'structure'),
-    lambda data, type_: type_.structure(data)  # type: ignore
+    lambda data, type_: getattr(type_, 'structure').__func__(type_, data)  # type: ignore
 )
 converter.register_unstructure_hook_func(  # type: ignore
     lambda obj: hasattr(obj, 'unstructure'),
     lambda obj: obj.unstructure()  # type: ignore
-)
-
-
-def structure_obj_to_generic_class(obj, cl):
-
-    def get_mapped_type(t, mapping):
-        if isinstance(t, typing.TypeVar):
-            return mapping.get(t.__name__, t)
-        try:
-            origin_type = t.__dict__['__origin__']
-            type_args = t.__dict__.get('__args__', [])
-            mapped_args = tuple(get_mapped_type(t_arg, mapping) for t_arg in type_args)
-
-            # TODO: support all generic types including custom generics
-            if origin_type == typing.Union:
-                return typing.Union[mapped_args]
-            if origin_type == dict:
-                return dict[mapped_args[0], mapped_args[1]]
-            if origin_type == list:
-                return list[mapped_args]
-            return t
-        except Exception as err:
-            return t
-
-    mapping = {}
-    if is_generic(cl):
-        base = get_origin(cl)
-        mapping = _generate_mapping(cl, mapping)
-        cl = base
-
-    for base in getattr(cl, "__orig_bases__", ()):
-        if is_generic(base) and not str(base).startswith("typing.Generic"):
-            mapping = _generate_mapping(base, mapping)
-            break
-    mapped_fields = []
-    for field in attr.fields(cl):
-        field_with_mapped_type = deepcopy(field).evolve(type=get_mapped_type(field.type, mapping))
-        mapped_fields.append(field_with_mapped_type)
-
-    cl_copy = type(f'{cl.__name__}_generated', cl.__bases__, dict(cl.__dict__))
-    cl_copy.__attrs_attrs__ = mapped_fields
-    return cl_copy.structure(obj)
-
-
-converter.register_structure_hook_func(
-    lambda type_: has_with_generic(type_) and hasattr(type_, 'structure'),
-    lambda data, type_: structure_obj_to_generic_class(data, type_)
 )
 
 converter.register_structure_hook(uuid.UUID, lambda data, type_: type_(data))  # type: ignore
